refactor(vendgingmachine.service.SettingMoneyService) : util에 있던 Money를 service로 이동
