+++ src/main/java/vendingmachine/service/MakingRandomCoinService.java
@@ -1,0 +1,4 @@
+package vendingmachine.service;
+
+public class MakingRandomCoinService {
+}
