+++ src/main/java/vendingmachine/service/OrderService.java
@@ -1,0 +1,4 @@
+package vendingmachine.service;
+
+public class OrderService {
+}
