+++ src/main/java/vendingmachine/service/SettingMoneyService.java
@@ -1,6 +1,8 @@
-package vendingmachine.util;
+package vendingmachine.service;
 
-public class Money {
+import vendingmachine.util.NumberException;
+
+public class SettingMoneyService {
 	public static int setMoney(String moneyStr) {
 		int money;
 		money = NumberException.checkMoneyException(moneyStr);
