+++ src/main/java/vendingmachine/service/VendingMachineService.java
@@ -1,0 +1,5 @@
+package vendingmachine.service;
+
+public class VendingMachineService {
+	
+}
